requirements.txt
Flask
gunicorn
nltk
pdfminer.six
tensorflow
